+++ src/examples/java/Main.java
@@ -1,4 +1,5 @@
 import com.github.johnnyjayjay.discord.commandapi.CommandSettings;
+import com.github.johnnyjayjay.discord.commandapi.DefaultHelpCommand;
 import net.dv8tion.jda.core.AccountType;
 import net.dv8tion.jda.core.JDA;
 import net.dv8tion.jda.core.JDABuilder;
@@ -14,19 +15,16 @@ public class Main {
     private static CommandSettings settings;
 
     public static void main(String[] args) throws LoginException, InterruptedException {
-        JDA jda = new JDABuilder(AccountType.BOT).setToken("VERY-SECRET-TOKEN-PLS-DONT-STEAL").buildBlocking();
+        JDA jda = new JDABuilder(AccountType.BOT).setToken("NDQwMjIyMTc1NjkwMDk2NjYx.DkZHIg.Fu6w20j9htc0Z_-G7W6S1k9Yumw").buildBlocking();
         // default prefix shall be "!" and we want the labels to be case insensitive.
         settings = new CommandSettings("!", jda, true);
         settings.setCooldown(3000) // commands can only be executed every 3 seconds now
-                .addHelpLabels("help", "helpme") // help can now be demanded by calling one of these labels.
+                //.addHelpLabels("help", "helpme") // help can now be demanded by calling one of these labels.
+                .put(new DefaultHelpCommand(), "help", "helpme")
                 .put(new PingCommand(), "ping")
                 .put(new PrefixCommand(settings), "setprefix")
                 .put(new KillCommand(settings), "kill", "deactivate", "shutdown") // setting aliases
                 .activate(); // Activating! Very important!
-    }
-
-    static class CooleClass {
-
     }
 
     public static String getPrefix(long guildId) {