+++ src/main/java/com/github/johnnyjayjay/discord/commandapi/CommandEvent.java
@@ -14,15 +14,18 @@ import java.util.Optional;
 /**
  * Represents a command event. This is not much different from a GuildMessageReceivedEvent, though it gives access to the called command.
  * @author Johnny_JayJay
- * @version 3.1_1
+ * @version 3.2
+ * @see GuildMessageReceivedEvent
  */
 public class CommandEvent extends GuildMessageReceivedEvent {
 
     private final Command command;
+    private final CommandSettings settings;
 
-    public CommandEvent(JDA api, long responseNumber, Message message, Command command) {
+    public CommandEvent(JDA api, long responseNumber, Message message, Command command, CommandSettings settings) {
         super(api, responseNumber, message);
         this.command = command;
+        this.settings = settings;
     }
 
     /**
@@ -58,6 +61,15 @@ public class CommandEvent extends GuildMessageReceivedEvent {
      */
     public Command getCommand() {
         return command;
+    }
+
+    /**
+     * Returns the CommandSettings instance this command was called for.
+     * @return an instance of CommandSettings.
+     * @see CommandSettings
+     */
+    public CommandSettings getCommandSettings() {
+        return settings;
     }
 
     /**